UefiCpuPkg/CpuDxe: fix an incorrect bit-wise operation
BZ: https://bugzilla.tianocore.org/show_bug.cgi?id=1164

The left operand is 64-bit but right operand could be 32-bit.
A typecast is a must because of '~' op before it.

Cc: Hao A Wu <hao.a.wu@intel.com>
Contributed-under: TianoCore Contribution Agreement 1.1

+++ UefiCpuPkg/CpuDxe/CpuPageTable.c
@@ -1181,7 +1181,7 @@ DebugExceptionHandler (
 
   for (PFEntry = 0; PFEntry < mPFEntryCount[CpuIndex]; PFEntry++) {
     if (mLastPFEntryPointer[CpuIndex][PFEntry] != NULL) {
-      *mLastPFEntryPointer[CpuIndex][PFEntry] &= ~IA32_PG_P;
+      *mLastPFEntryPointer[CpuIndex][PFEntry] &= ~(UINT64)IA32_PG_P;
     }
   }
